Some cleanup. Added Fox alerts to Camera group.
