+++ EnchantedGarden/Assets/Prefabs/Systems/Multi-Target Camera.prefab
@@ -289,7 +289,7 @@ GameObject:
   m_Component:
   - component: {fileID: 4488676007313481970}
   m_Layer: 0
-  m_Name: Follow Player and Spirits Camera
+  m_Name: Multi-Target Camera
   m_TagString: Untagged
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
